requirements.txt
# 核心依赖
openai==1.54.2
httpx==0.27.2
anthropic==0.39.0
google-generativeai==0.8.4
google-genai==1.7.0
python-dotenv==1.0.0

# interface
fastapi==0.115.5
uvicorn==0.32.1

# 测试依赖
pytest==8.0.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# 开发工具
black==23.9.1
isort==5.12.0

# 可选依赖
uuid==1.30
Pillow==11.1.0
anthropic[bedrock]
boto3
